Add files via upload
This is a frog leap game. I have written a code in Python on how this game works. The game is to interchange the places of green color frogs with brown color frogs.
I have used jupyter notebook to write the code and used infinite loop.
There are few conditions in the game like, when the user gives certain values for input - it gets invalid and the loop should start from the beginning again. And if the user enters 'q', the game should end which means the user wants to quit. 
And when the user provides valid input, the frog place will be gradually changing and the required order should be attained. When the required order is attained the game ends and the user wins the game.